Unir los 3 ViewModels anteriores de DetailedMetrics en uno solo - DetailedMetricsViewModel, modificación de MAIN en Manifest.
