+++ app/src/main/AndroidManifest.xml
@@ -33,10 +33,6 @@
         <activity
             android:name=".framework.view.SupportActivity"
             android:exported="true">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
         </activity>
         <activity
             android:name="com.journeyapps.barcodescanner.CaptureActivity"
@@ -138,6 +134,10 @@
         <activity
             android:name=".framework.view.DetailedMetrics"
             android:exported="true">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
         </activity>
         <activity
             android:name=".framework.view.CheckIfLogged"
